requirements.txt
pymysensors==0.10.0
